Update of ApplicationDbContect relationships

+++ Models/Order.cs
@@ -17,7 +17,7 @@ namespace GeorgianGroceries.Models
         public string Province { get; set; }
         public string PostalCode { get; set; }
         public string Phone { get; set; }
-        public Decimal Total { get; set; }
+        public float Total { get; set; }
         //Child reference 
         public List<OrderDetail> OrderDetails { get; set; }
     }
